Rename enhancements.color_mapping to enhancements.colormap
