+++ satpy/enhancements/colormap.py
