+++ satpy/etc/enhancements/glm.yaml
@@ -3,7 +3,7 @@ enhancements:
     name: flash_extent_density
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - {colors: ylorrd, min_value: 0, max_value: 20}
