+++ satpy/etc/enhancements/li.yaml
@@ -6,7 +6,7 @@ enhancements:
     standard_name: acc_flash
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - { colors: ylorrd, min_value: 0, max_value: 5 }
@@ -15,7 +15,7 @@ enhancements:
     standard_name: acc_flash_alpha
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - {
@@ -30,7 +30,7 @@ enhancements:
     standard_name: acc_flash_area
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - { colors: ylorrd, min_value: 0, max_value: 20 }
@@ -39,7 +39,7 @@ enhancements:
     standard_name: acc_flash_area_alpha
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - {
@@ -54,7 +54,7 @@ enhancements:
     standard_name: lightning_radiance
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - { colors: ylorrd, min_value: 0, max_value: 1000 }
@@ -63,7 +63,7 @@ enhancements:
     standard_name: lightning_radiance_alpha
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - {
@@ -78,7 +78,7 @@ enhancements:
     standard_name: lightning_time
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - { colors: ylorrd, reverse: True, min_value: 0, max_value: 1 }
