+++ satpy/etc/enhancements/scatterometer.yaml
@@ -4,7 +4,7 @@ enhancements:
     name: scat_wind_speed
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - {colors: spectral, min_value: 0, max_value: 25}
@@ -13,7 +13,7 @@ enhancements:
     standard_name: soil_moisture
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - {colors: rdbu, min_value: 0, max_value: 100}
