+++ satpy/etc/enhancements/viirs.yaml
@@ -84,7 +84,7 @@ enhancements:
     standard_name: viirs_flames_colorised
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - { colors: ylorrd, min_value: 330, max_value: 430 , reverse: true}
@@ -94,7 +94,7 @@ enhancements:
     standard_name: night_ir105
     operations:
       - name: colorize
-        method: !!python/name:satpy.enhancements.color_mapping.colorize
+        method: !!python/name:satpy.enhancements.colormap.colorize
         kwargs:
           palettes:
             - { colors: greys, min_value: 190, max_value: 295 }
